Packages Lockfile
Going to look into modernizing this project! I looked up to it a bunch when I started adding file system support to https://stedit.app

It's nice that when I first saw this project a few years ago, it was too complex for me to understand the tooling, and now it doesn't seem so far-fetched. I think moving it over to something like Vite will work nicely! Same with TypeScript and ESM.