Full JSDoc Type Safety!
Yeah, this is definitely one of the easier ways to migrate untyped projects over to types! It works great! It allows you to safely document the code, without needing to add in a new build tool or build config step. I think that's essential to ensuring that the code stays the same while you document it. In-place documentation, if it were!