Vite Setup!
I still have to properly add Vite as a dependency (the current dependency issues don't allow for it to run), and I have to move the Gulpfile configurations over to Vite, like Workbox for example. Really happy to see the app does actually start though! You just have to navigate to the `src/` folder when running the dev server.

Can't believe the FSAAPI has been 'around' for 4 years now! It's only been stable for a few of those, but the fact that it's even in stable nowadays is amazing and really cool. Still only mainly Chromium browsers support it though.