ESM Build Target
This helps with properly setting up the configuration for the project. Not having NodeNext as the module setup was definitely hindering things from being declared properly. I still have to audit the use of packages here though, as that's what seems to be breaking the Vite dev server when running dev mode.