Indentation Preservation
Migrating from HTML to JSX has some challenges, mostly in places where HTML's formatting is used to dictate where whitespace should be in the page. JSX doesn't persist that losslessly quite right, so you have to escape it with a string literal and brackets. I ran into this same issue with Dovetail too.