TextArea Component
Realized I have to lazy load the components still, because loading them with imperative imports changes the order around as to what will resolve first. I want to, and will be able to make them proper imperative imports again, once the state can be demanged, which can be done once the parts of the JSX are put into their own components again.