MenuEdit Component
Time Travel in Texas kind of reminds me of Opeth too. It would be neat to hear a crossover between Opeth and no-man.

The woodwind sample sounds to be the same one from the bonus disc for Signify too, that's another cool crossover as well. There's the reverse sound from Yellow Hedgerow too. Wait, maybe it's Up the Downstair? It might be the Infinities bonus disc from that instead.

Oh yeah and My Revenge on Seattle reminds me of Voyage 34. Man, no-man is awesome! It sounds like some people weren't sure about this album too. It has a huge bunch of stylistic crossovers that are some of my favorites, from Steven & friends.