Declarative Component Imports
Had to debug the import stack order, it was getting stuck importing things between the menus module, and the app module. I fixed it with dynamically importing the app module inside the menu module.

Later DTP actually has hinting sounds that point forward to the sound that Empath has, very interesting!

I was listening to Wild Opera, and one of the songs reminded me of Holy Drinker from Raven as well.

https://www.youtube.com/watch?v=n-DKs0qfdEk
https://www.youtube.com/watch?v=YhcoLO8vZZU
https://www.youtube.com/watch?v=52MiJ1yn8jE
https://www.youtube.com/watch?v=KILt_krL5iU
https://www.youtube.com/watch?v=Yv9iR1K9UmM